Using prepared RoninOS repo from '/tmp/overlay/' instead of cloning
The readme instructs to clone the RoninOS repo already and to prepare it for the Armbian build process in the 'userpatches/overlay' directory. So the user could have made edits to it. These 'overlay' files are later available in '/tmp/overlay/' during build time to make use of them. No need to rashly clone again and throw away possible user-manufactured changes.